+++ Tests/Integration/MSFT_SqlSetup.config.ps1
@@ -95,17 +95,17 @@ Configuration MSFT_SqlSetup_CreateDependencies_Config
     )
 
     Import-DscResource -ModuleName 'PSDscResources'
-    Import-DscResource -ModuleName 'xStorage'
+    Import-DscResource -ModuleName 'StorageDsc'
 
     node localhost {
-        xMountImage 'MountIsoMedia'
+        MountImage 'MountIsoMedia'
         {
             ImagePath   = $Node.ImagePath
             DriveLetter = $Node.DriveLetter
             Ensure      = 'Present'
         }
 
-        xWaitForVolume WaitForMountOfIsoMedia
+        WaitForVolume WaitForMountOfIsoMedia
         {
             DriveLetter      = $Node.DriveLetter
             RetryIntervalSec = 5